refactor(app): query structure
