+++ app/src/lib/graphql/queries/faucet.ts
