+++ app/src/lib/graphql/queries/index-status.ts
