+++ app/src/lib/graphql/queries/stats.ts
