+++ app/src/routes/faucet/(components)/dydx-faucet.svelte
@@ -11,7 +11,7 @@ import { Button } from "$lib/components/ui/button/index.ts"
 import SpinnerSVG from "$lib/components/spinner-svg.svelte"
 import { cosmosStore } from "$/lib/wallet/cosmos/config.ts"
 import { derived, writable, type Writable } from "svelte/store"
-import { dydxFaucetMutation } from "$lib/graphql/documents/faucet"
+import { dydxFaucetMutation } from "$lib/graphql/queries/faucet"
 import { getCosmosChainBalances } from "$lib/queries/balance/cosmos"
 import { isValidCosmosAddress } from "$lib/wallet/utilities/validate.ts"
 import type { AwaitedReturnType, DiscriminatedUnion } from "$lib/utilities/types.ts"
